# [FoldConstant] Create Interpreter for each constant subgraph