This is related to https://discuss.tvm.ai/t/vm-slow-compilation-of-tf-object-detection-models/7479

For TF object detection models, a module has many functions (TF SSD mobilenet has 48), where a couple of functions are pretty huge (more than 10k call nodes). FoldConstant, because it is a function pass, is called for each function in the module. However, FoldConstant also creates an Interpreter on every invocation, which currently is based on the full mod and therefore expensive.

This PR creates an interpreter for each constant subgraph. I am not sure if this is the right way. The purpose of this PR is to start a discussion and identify if there is some other higher-level design issue that needs to be resolved.

With this PR, compilation time of
* SSD reduced from 955 to 453 seconds
* Faster RCNN reduced from 4630 to 1227 seconds

@zhiics @masahi @kevinthesun @icemelon9